refactor(errorCalculation): use linear function instead of sinusoid for test
Although the Levenberg-Marquardt algorithm is generally used to fit non-linear functions, there is
no need to use a non-linear function in the unit test suite for the residual/error calculation code.
That computation is currently a simple sum of (absolute) differences, which can be tested with
simpler functions (even a constant one). This commit demonstrates that there is no reduction in code
coverage, yet the test case is easier to reason about.