fix: throw error if initialValues missing (fix #52)
Previously, the code attempted to use the following expression
for initialValues if it was not provided:
`new Array(parameterizedFunction.length).fill(1);`

Unfortunately, since the parameterized function always takes a single array as its
argument, accessing `.length` always returns 1, not the number of parameters.
The code now throws an error if initialValues is not provided rather than silently failing.
The initialValues option is specified as mandatory so this is not considered a breaking change.

This commit also includes some minor typographical improvements to doc comments.